Sort modules structure
